+++ ts/src/routes/help/+page.svelte
@@ -1,90 +1,100 @@
 <script lang="ts">
-    import {
-        client,
-        type GetLanguagesAndProvidersResponse,
-        type Provider,
-    } from "$lib";
-    import Error from "$lib/Error.svelte";
-    import LanguageSelect from "$lib/LanguageSelect.svelte";
-    // import { type GetLanguagesAndProvidersResponse, type Provider } from "$lib/generated/backend_pb";
-    import SelectControl from "$lib/SelectControl.svelte";
-    import { promiseWithResolver, type SelectOption, Spinner } from "ankiutils";
+    import { client, type GetSupportLinksResponse } from "$lib";
+    import { promiseWithResolver, Spinner } from "ankiutils";
+
     import { onMount } from "svelte";
 
-    let [languagesPromise, resolveLanguages, _] = promiseWithResolver<
-        GetLanguagesAndProvidersResponse
+    let [linksPromise, resolveLinks] = promiseWithResolver<
+        GetSupportLinksResponse
     >();
-    let language = $state("eng");
-    let languages = $state<SelectOption[]>([]);
-    let providers = $state<Provider[]>([]);
 
-    async function onLanguageSelected(lang: string) {
-        console.log("onLanguageSelected", lang, language);
-        const response = await client.getProvidersForLanguage({
-            language: lang,
-        });
-        providers = response.providers;
-    }
-
-    onMount(async () => {
-        const response = await client.getLanguagesAndProviders({
-            defaultLanguage: language,
-        });
-        resolveLanguages(response);
-        languages = response.languages.map(lang => {
-            return { label: lang.name, value: lang.code };
-        });
-        providers = response.providers;
+    onMount(() => {
+        client.getSupportLinks({}).then(resolveLinks);
     });
 </script>
 
-<div class="container mx-auto min-h-screen flex flex-col gap-4 p-4">
-    <h2 class="font-bold text-4xl my-4">Supported Languages and Providers</h2>
-
-    <p>
-        This page lists all languages supported by InContext, along with the
-        providers available for each language. The codes are useful if you're
-        using the add-on's custom template filter.
-    </p>
-
-    {#await languagesPromise}
-        <Spinner label="Loading languages..." />
-    {:then _}
-        <SelectControl label="Language">
-            <LanguageSelect
-                options={languages}
-                onSelected={onLanguageSelected}
-                bind:value={language}
-            />
-        </SelectControl>
-        <p>Language code: {language}</p>
-        <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
-            {#each providers as provider (provider.code)}
-                <div class="card bg-base-100 shadow-sm">
-                    <div class="card-body">
-                        <h2 class="card-title">{provider.name}</h2>
-                        <p>
-                            code: {provider.code}
-                        </p>
-                        <div class="card-actions justify-end">
-                            <a
-                                class="btn btn-primary"
-                                href={provider.url}
-                                target="_blank"
-                                aria-label="Visit website"
-                            >Visit</a>
-                        </div>
-                    </div>
-                </div>
-            {/each}
-        </div>
-    {:catch error}
-        <Error error={error.rawMessage} />
+<div class="m-4 p-4 card w-min-screen shadow-lg bg-slate-200">
+    {#await linksPromise}
+        <Spinner />
+    {:then response}
+        <main class="prose prose-2xl prose-h2:text-4xl">
+            <h2>Usage</h2>
+            <p>
+                See the <a href={response.docsPage} target="_blank">docs</a> for
+                usage.
+            </p>
+            <h2>Changelog</h2>
+            <p>
+                See <a
+                    href="{response.githubPage}/blob/main/CHANGELOG.md"
+                    target="_blank"
+                >CHANGELOG.md</a> for a list of changes.
+            </p>
+            <h2>Support & feature requests</h2>
+            <p>
+                Please post any questions, bug reports, or feature requests in
+                the
+                <a href={response.forumsPage} target="_blank">support page</a>
+                or the <a
+                    href="{response.githubPage}/issues"
+                    target="_blank"
+                >issue tracker</a>.
+            </p>
+            <p>
+                If you want priority support for your feature/help request, I'm
+                available for hire. Get in touch via <a
+                    href="mailto:abdo@abdnh.com"
+                    target="_blank"
+                >email</a> or the Upwork link below.
+            </p>
+            <h2>Support me</h2>
+            <p>Consider supporting me if you like my work:</p>
+            <a
+                href="https://github.com/sponsors/abdnh"
+                target="_blank"
+                aria-label="GitHub Sponsors link"
+            ><img
+                    src="/github.png"
+                    alt="GitHub logo"
+                    class="not-prose"
+                ></a>
+            <a
+                href="https://www.patreon.com/abdnh"
+                target="_blank"
+                aria-label="Patreon link"
+            ><img
+                    src="/patreon.png"
+                    alt="Patreon"
+                    class="not-prose"
+                ></a>
+            <a
+                href="https://www.buymeacoffee.com/abdnh"
+                target="_blank"
+                aria-label="Buy Me a Coffee link"
+            ><img
+                    src="/buymeacoffee.png"
+                    alt="Buy Me A Coffee"
+                    class="not-prose"
+                ></a>
+            <p>I'm also available for freelance add-on development:</p>
+            <a
+                href="https://www.upwork.com/freelancers/~01d764ac58a0eccc5c"
+                target="_blank"
+                aria-label="Upwork link"
+            ><img
+                    src="/upwork.png"
+                    alt="Upwork"
+                    class="not-prose"
+                ></a>
+        </main>
     {/await}
 </div>
 
-<style lang="scss">
-    :global(html) {
-        font-size: 20px;
+<style>
+    @reference "tailwindcss";
+
+    img {
+        @apply h-20;
+        display: inline-block;
     }
 </style>